Moved to postgres from json and added translation

+++ commands/MessageFilter/Filter.ts
@@ -1,10 +1,11 @@
 import discord from "discord.js";
 import Filter from "bad-words";
-import { Json } from "../../util/Json";
+import { db } from "../../db/db";
+import { Tranlator } from "../../Translator/Translator";
 
 export class MessageFilter {
   public static filter = new Filter();
-  private curseCount: { username: string; word: string }[] = [];
+  private curseCount: IHistoryEntry[] = [];
 
   constructor(cli: discord.Client) {
     // * add proper data santization
@@ -12,9 +13,13 @@ export class MessageFilter {
 
     this.Init();
 
-    cli.on("message", (message) => {
+    cli.on("messageUpdate", async (_, message) => {
+      if (message.content == null || message.author == null) return;
+
       const author = message.author.username;
       const content = message.content.toLowerCase();
+
+      const history = await db.blocklist.History.Get(author);
 
       var count: number = 0;
 
@@ -24,11 +29,7 @@ export class MessageFilter {
       if (content.includes("#deleteword")) return;
 
       if (MessageFilter.filter.isProfane(content)) {
-        const file = JSON.parse(Json.Read("config.json").toString());
-
-        this.curseCount = file.users;
-
-        this.curseCount.forEach((user) => {
+        history.forEach((user) => {
           if (user.username == author) {
             count++;
           }
@@ -48,12 +49,48 @@ export class MessageFilter {
 
         this.curseCount.push({
           username: author,
-          word: content,
+          message: content,
         });
 
-        file.users = this.curseCount;
+        if (message.deletable) {
+          message.delete();
+        }
+      }
+    });
 
-        Json.Write("config.json", JSON.stringify(file));
+    cli.on("message", async (message) => {
+      if (message.content == null || message.author == null) return;
+
+      const author = message.author.username;
+      const content = await Tranlator.en.Translate(
+        message.content.toLowerCase()
+      );
+
+      const history = await db.blocklist.History.Get(author);
+
+      var count: number = 0;
+
+      if (author == "Pinky" || author == "Pinky Dev") return;
+
+      if (content.includes("#addword")) return;
+      if (content.includes("#deleteword")) return;
+
+      if (MessageFilter.filter.isProfane(content)) {
+        history.forEach(() => count++);
+
+        if (count > 10) {
+          message.reply(
+            `you know im going to delete your messages anyway :rolling_eyes:`
+          );
+        } else if (count > 5) {
+          message.reply(
+            `HOW MANY TIMES DO I HAVE TO TELL YOU STOP CURSING!!!!!`
+          );
+        } else if (count < 5) {
+          message.reply(`stop cursing`);
+        }
+
+        await db.blocklist.History.Add(content, author);
 
         if (message.deletable) {
           message.delete();
@@ -62,9 +99,9 @@ export class MessageFilter {
     });
   }
 
-  private Init() {
-    const file = JSON.parse(Json.Read("config.json").toString());
+  private async Init() {
+    const list = await db.blocklist.Get();
 
-    file.words.forEach((word: string) => MessageFilter.filter.addWords(word));
+    list.forEach((l) => MessageFilter.filter.addWords(l.word));
   }
 }
